`use Md.Parser` works (no opts yet)
